fix(ui): move offline indicator from top banner to bottom toast (#62)
* fix(ui): move offline indicator from top banner to bottom toast

Replaces the in-content .offline-badge inside Home.svelte with a
persistent bottom-pinned .offline-toast in App.svelte. The new
indicator matches the visual language of the existing error toast
(same position, same danger color vars, same shadow) but has no
dismiss button and stays until connectivity returns. The transient
error toast shifts up to 66px when both are visible to prevent
overlap. Removes all dead CSS from Home.svelte (.offline-badge,
.offline-dot, dot-pulse keyframes, .update-banner.dimmed).


* fix(offline-toast): move connectivity polling to App.svelte and fix fly/transform conflict

Polling was confined to Home.svelte so the isOnline store stopped updating
when the user navigated to other views. Moves pingConnectivity + the
visibilitychange-aware interval to App.svelte so it runs for the lifetime
of the app.

Also fixes the transition:fly + transform:translateX(-50%) centering
conflict on .offline-toast: Svelte's fly applies an inline transform that
overwrites the CSS centering transform, causing a horizontal jump during
the slide animation. Switches to margin-inline:auto / width:fit-content
centering so fly can animate without interference.


* fix(offline-toast): prevent interval leak in visibilitychange handler

Always clear the existing connectivity interval before starting a new one.
Without this, a redundant 'visible' event (possible on some browsers)
creates a second interval while the first keeps running, leaking timer
references indefinitely.


---------
